Create report function
This function is giving me some trouble, I've sorted the dict properly
highest to lowest but to list its key and value with a string of text
making a pretty report is giving me some issues. It has to do with
accesing the key values pairs differently once its been conveerted
to a list for sorting. Will keep trying.